backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
psycopg[binary]==3.2.1
alembic==1.13.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic-settings==2.5.2
python-multipart==0.0.9
redis==5.0.8
celery==5.4.0
httpx==0.27.2
beautifulsoup4==4.12.3
